relayburn: mechanical Rust cleanup (clippy + dedup + dead code)
Pure cleanup, no behavior change. Drops the workspace clippy warning
count from 28 to 17 and removes ~90 LOC of duplication and dead code.

- relayburn-cli: replace `io::Error::new(io::ErrorKind::Other, e)` with
  `io::Error::other(e)` (×6 in render/error.rs, render/json.rs,
  commands/overhead.rs).
- relayburn-cli: collapse redundant `Ok(Ledger::open(opts)?)` to plain
  `Ledger::open(opts)` in commands/{ingest,mcp_server}.rs (the SDK verb
  already returns `anyhow::Result`).
- relayburn-cli: `Some(n) if n == 0` → `Some(0)`, `% 3 == 0` →
  `.is_multiple_of(3)` in commands/{ingest,overhead}.rs.
- relayburn-sdk: `t.files_touched.as_ref().map(|v| v.as_slice())` →
  `.as_deref()` in reader/claude.rs.
- relayburn-sdk: consolidate four byte-identical copies of
  `format_with_commas` (findings, tool_call_patterns, tool_output_bloat,
  ghost_surface) into a new `analyze::util` module.
- relayburn-sdk: delete `lookup_content_fingerprint` from
  ledger/reader.rs — `#[allow(dead_code)]` and confirmed unused.
- relayburn-cli: delete `not_yet_implemented` from commands/mod.rs —
  `#[allow(dead_code)]` and not called now that all Wave 2 D1–D8
  presenters have landed.
- relayburn-sdk: drop incorrect `#[allow(dead_code)]` on
  `writer::debug_now` — it IS called from ledger.rs (twice).
- relayburn-cli: drop the stale `#[allow(dead_code)]` on `GlobalArgs`
  now that Wave 2 consumes both `ledger_path` and `no_color`.

This is the lowest-risk slice of a larger Rust review punch list
(perf wins, async hygiene, SQL filter pushdown, env-var removal, etc.) —
those land separately.
